+++ noodles-bgzf/src/gzi/async/fs.rs
@@ -4,10 +4,10 @@ use std::path::Path;
 
 use tokio::{
     fs::File,
-    io::{self, BufReader},
+    io::{self, AsyncWriteExt, BufReader, BufWriter},
 };
 
-use super::io::Reader;
+use super::io::{Reader, Writer};
 use crate::gzi::Index;
 
 /// Reads the entire contents of a GZ index.
@@ -31,4 +31,36 @@ where
 {
     let mut reader = File::open(src).await.map(BufReader::new).map(Reader::new)?;
     reader.read_index().await
+}
+
+/// Writes a GZ index to a file.
+///
+/// This is a convenience function and is equivalent to creating a file at the given path and
+/// writing the index.
+///
+/// # Examples
+///
+/// ```no_run
+/// # #[tokio::main]
+/// # async fn main() -> tokio::io::Result<()> {
+/// use noodles_bgzf::gzi;
+/// let index = gzi::Index::from(vec![(0, 0)]);
+/// gzi::r#async::fs::write("in.gz.gzi", &index).await?;
+/// # Ok(())
+/// # }
+/// ```
+pub async fn write<P>(dst: P, index: &Index) -> io::Result<()>
+where
+    P: AsRef<Path>,
+{
+    let mut writer = File::create(dst)
+        .await
+        .map(BufWriter::new)
+        .map(Writer::new)?;
+
+    writer.write_index(index).await?;
+
+    writer.get_mut().shutdown().await?;
+
+    Ok(())
 }